refactor: Player의 일급 컬렉션 Players 제거 후 역할을Participants에 위임
