--- src/test/java/domain/PlayersTest.java
@@ -1,51 +1,0 @@
-package domain;
-
-import domain.participant.Players;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-
-class PlayersTest {
-
-    @ParameterizedTest
-    @ValueSource(strings = {"가비", "연어,가비,애쉬,우가,럿고,비버,아코"})
-    void 플레이어_수는_1에서_7명이다(String names) {
-        //given
-
-        //when
-
-        //then
-        assertDoesNotThrow(() -> Players.of(List.of(names.split(","))));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {",", "연어,가비,애쉬,우가,럿고,비버,아코,네오"})
-    void 플레이어_수는_1에서_7명이_아니면_예외가_발생한다(String names) {
-        //given
-
-        //when
-
-        //then
-        assertThatThrownBy(() -> Players.of(List.of(names.split(","))))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 플레이어의 수는 1 ~ 7 이내여야 합니다");
-    }
-
-    @Test
-    void 중복된_이름을_입력하면_예외가_발생한다() {
-        //given
-        List<String> names = List.of("도비", "   도비");
-
-        //when
-
-        //then
-        assertThatThrownBy(() -> Players.of(names))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 플레이어의 이름은 중복될 수 없습니다");
-    }
-}
