+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
-import domain.participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
